Type: move validity pattern to another file
because timezone_handler needs it
