+++ source/ed/connectors/gtfs_parser.cpp
@@ -737,7 +737,7 @@ void CalendarGtfsHandler::finish(Data& data) {
     BOOST_ASSERT(data.validity_patterns.size() == gtfs_data.tz.vp_by_name.size());
 }
 
-void CalendarGtfsHandler::handle_line(Data&, const csv_row& row, bool) {
+void CalendarGtfsHandler::handle_line(Data& data, const csv_row& row, bool) {
     if (row.empty())
         return;
     std::bitset<7> week;
@@ -747,8 +747,8 @@ void CalendarGtfsHandler::handle_line(Data&, const csv_row& row, bool) {
         return;
     }
     //we only build temporary validity pattern that will be split after the CalendarDateParser has been called
-    nm::ValidityPattern& vp = gtfs_data.tz.non_split_vp[row[id_c]];
-    vp.beginning_date = gtfs_data.production_date.begin();
+    nt::ValidityPattern& vp = gtfs_data.tz.non_split_vp[row[id_c]];
+    vp.beginning_date = data.meta.production_date.begin();
     vp.uri = row[id_c];
 
     //week day init (remember that sunday is the first day of week in the us)
@@ -762,13 +762,16 @@ void CalendarGtfsHandler::handle_line(Data&, const csv_row& row, bool) {
 
     //Init the validity period
     boost::gregorian::date b_date = boost::gregorian::from_undelimited_string(row[start_date_c]);
-    boost::gregorian::date_period period = boost::gregorian::date_period(
-                (b_date > gtfs_data.production_date.begin() ? b_date : gtfs_data.production_date.begin()), boost::gregorian::from_undelimited_string(row[end_date_c]) + boost::gregorian::days(1));
+
     //we add one day to the last day of the period because end_date_c is included in the period
     //and as say the constructor of boost periods, the second args is "end" so it is not included
+    boost::gregorian::date e_date = boost::gregorian::from_undelimited_string(row[end_date_c]) + 1_days;
 
-    for(boost::gregorian::day_iterator it(period.begin()); it<period.end(); ++it) {
-        if(week.test((*it).day_of_week())) {
+    boost::gregorian::date_period full_period (b_date, e_date);
+    auto period = full_period.intersection(data.meta.production_date);
+
+    for (boost::gregorian::day_iterator it(period.begin()); it<period.end(); ++it) {
+        if (week.test((*it).day_of_week())) {
             vp.add((*it));
         } else {
             vp.remove((*it));
@@ -776,7 +779,7 @@ void CalendarGtfsHandler::handle_line(Data&, const csv_row& row, bool) {
     }
 }
 
-static boost::gregorian::date_period compute_smallest_active_period(const nm::ValidityPattern& vp) {
+static boost::gregorian::date_period compute_smallest_active_period(const nt::ValidityPattern& vp) {
     size_t beg(std::numeric_limits<size_t>::max()), end(std::numeric_limits<size_t>::min());
 
     for (size_t i = 0; i < vp.days.size(); ++i) {
@@ -813,7 +816,7 @@ void split_validity_pattern_over_dst(Data& data, GtfsData& gtfs_data) {
     data.tz_handler = nt::TimeZoneHandler(data.meta, splited_production_period);
 
     for (const auto& name_and_vp: gtfs_data.tz.non_split_vp) {
-        const nm::ValidityPattern& original_vp = name_and_vp.second;
+        const nt::ValidityPattern& original_vp = name_and_vp.second;
 
         boost::gregorian::date_period smallest_active_period = compute_smallest_active_period(original_vp);
 
@@ -823,7 +826,7 @@ void split_validity_pattern_over_dst(Data& data, GtfsData& gtfs_data) {
 
         size_t cpt(1);
         for (const auto& utc_shit_and_periods: split_periods) {
-            nm::ValidityPattern* vp = new nm::ValidityPattern(gtfs_data.production_date.begin());
+            nt::ValidityPattern* vp = new nt::ValidityPattern(data.meta.production_date.begin());
 
             for (const auto& period: utc_shit_and_periods.second) {
                 for (boost::gregorian::day_iterator it(period.begin()); it < period.end(); ++it) {
@@ -858,16 +861,16 @@ void CalendarDatesGtfsHandler::init(Data&) {
             e_type_c = csv.get_pos_col("exception_type");
 }
 
-void CalendarDatesGtfsHandler::handle_line(Data&, const csv_row& row, bool) {
+void CalendarDatesGtfsHandler::handle_line(Data& data, const csv_row& row, bool) {
     if(row.empty())
         return;
 
     bool new_vp = gtfs_data.tz.non_split_vp.find(row[id_c]) == gtfs_data.tz.non_split_vp.end();
 
-    nm::ValidityPattern& vp = gtfs_data.tz.non_split_vp[row[id_c]];
+    nt::ValidityPattern& vp = gtfs_data.tz.non_split_vp[row[id_c]];
 
     if (new_vp) {
-        vp.beginning_date = gtfs_data.production_date.begin();
+        vp.beginning_date = data.meta.production_date.begin();
         vp.uri = row[id_c];
     }
 
@@ -993,7 +996,7 @@ void TripsGtfsHandler::handle_line(Data& data, const csv_row& row, bool) {
     //the validity pattern may have been split because of DST, so we need to create one vj for each
     for (auto vp_it = vp_range.first; vp_it != vp_range.second; ++vp_it, cpt_vj++) {
 
-        nm::ValidityPattern* vp_xx = vp_it->second;
+        nt::ValidityPattern* vp_xx = vp_it->second;
 
         nm::VehicleJourney* vj = new nm::VehicleJourney();
         const std::string original_uri = row[trip_c];
@@ -1115,7 +1118,7 @@ std::vector<nm::StopTime*> StopTimeGtfsHandler::handle_line(Data& data, const cs
         nm::StopTime* stop_time = new nm::StopTime();
 
         //we need to convert the stop times in UTC
-        int utc_offset = data.tz_handler.get_utc_offset(vj_it->second->validity_pattern);
+        int utc_offset = data.tz_handler.get_first_utc_offset(*vj_it->second->validity_pattern);
 
         stop_time->arrival_time = to_utc(row[arrival_c], utc_offset);
         stop_time->departure_time = to_utc(row[departure_c], utc_offset);
@@ -1165,7 +1168,7 @@ void FrequenciesGtfsHandler::handle_line(Data& data, const csv_row& row, bool) {
         }
 
         //we need to convert the stop times in UTC
-        int utc_offset = data.tz_handler.get_utc_offset(vj_it->second->validity_pattern);
+        int utc_offset = data.tz_handler.get_first_utc_offset(*vj_it->second->validity_pattern);
 
         vj_it->second->start_time = to_utc(row[start_time_c], utc_offset);
         vj_it->second->end_time = to_utc(row[end_time_c], utc_offset);
@@ -1197,7 +1200,6 @@ void GenericGtfsParser::fill(Data& data, const std::string& beginning_date) {
         LOG4CPLUS_FATAL(logger, error);
         throw navitia::exception(error);
     }
-    data.meta.production_date = gtfs_data.production_date;
     data.meta.timezone = gtfs_data.tz.default_timezone.first;
 }
 
@@ -1405,7 +1407,8 @@ void GenericGtfsParser::manage_production_date(Data& data, const std::string& be
     auto end_it = data.feed_infos.find("feed_end_date");
     boost::gregorian::date start_date(boost::gregorian::not_a_date_time),
             end_date(boost::gregorian::not_a_date_time);
-    if ((start_it != data.feed_infos.end()) && (end_it != data.feed_infos.end())) {
+    if ((start_it != data.feed_infos.end() && start_it->second != "")
+            && (end_it != data.feed_infos.end() && end_it->second != "")) {
         try{
             start_date = boost::gregorian::from_undelimited_string(start_it->second);
         }catch(const std::exception& e) {
@@ -1423,14 +1426,14 @@ void GenericGtfsParser::manage_production_date(Data& data, const std::string& be
     }
     if ((start_date.is_not_a_date()) || (end_date.is_not_a_date())){
         try {
-            gtfs_data.production_date = find_production_date(beginning_date);
+            data.meta.production_date = find_production_date(beginning_date);
         } catch (...) {
             if(beginning_date == "") {
                 LOG4CPLUS_FATAL(logger, "Impossible to find the production date");
             }
         }
     } else {
-        gtfs_data.production_date = complete_production_date(beginning_date, start_date, end_date);
+        data.meta.production_date = complete_production_date(beginning_date, start_date, end_date);
     }
 }
